requirements.txt
PyYAML~=6.0
avionix~=0.4.5
Flask~=2.0.3
GitPython~=3.1.27
